Finished splitting IVPs
